refatoração & isolamento de responsabilidade
